Relocate PayConsumerProduct prsntr + cntrllr
